% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield_histogram.R
\name{datashield_histogram}
\alias{datashield_histogram}
\title{DataSHIELD helper function for histograms}
\usage{
datashield_histogram(opal_connection, method = "deterministic")
}
\arguments{
\item{opal_connection}{An Opal connection}

\item{method}{}
}
\value{
It creates a directory with subdirectories of each study, containing histograms of all variables
}
\description{
DataSHIELD helper function for histograms
}
\examples{
datashield_histogram(opal_connection = opal_connection)
}
