% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield_table.R
\name{datashield_table}
\alias{datashield_table}
\title{Table function for categorical data}
\usage{
datashield_table(opal_connection)
}
\arguments{
\item{opal_connection}{}
}
\value{
runs the datashield function ds.table() on all categorical variables of all studies and outputs a table with the outcome
}
\description{
Table function for categorical data
}
\examples{
ds.table(opals)
}
