% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.histAllVars.R
\name{ds.histAllVars}
\alias{ds.histAllVars}
\title{DataSHIELD helper function for histograms}
\usage{
ds.histAllVars(datasources = NULL, method = "deterministic")
}
\arguments{
\item{datasources}{An Opal connection}

\item{method}{Type of method for building the histogram}
}
\value{
It creates a directory with subdirectories of each study, containing histograms of all variables
}
\description{
DataSHIELD helper function for histograms
}
