% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.rse.R
\name{ds.rse}
\alias{ds.rse}
\title{Robust Standard Errors HC1}
\usage{
ds.rse(formula, datasources, type = "split", data)
}
\arguments{
\item{formula}{The formula to perform RSE e.g. "Age ~ Sex + Weight"}

\item{datasources}{A list of \code{\link{DSConnection-class}} objects obtained after login.
If the \code{datasources} argument is not specified the default set of connections will be
used: see \code{\link{datashield.connections_default}}.}

\item{type}{"split" or "combine"}

\item{data}{A data.frame on the server-side}
}
\value{
robust standard errors of the formula
}
\description{
Robust Standard Errors HC1
}
