% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tableBatch.R
\name{ds.tableBatch}
\alias{ds.tableBatch}
\title{Table function for categorical data}
\usage{
ds.tableBatch(df = "D", datasources = NULL)
}
\arguments{
\item{df}{specifies the  df that was assigned in the login, default is "D"}

\item{datasources}{An opal connection}
}
\value{
runs the datashield function ds.table on all categorical variables of all studies and outputs a table with the outcome
}
\description{
Table function for categorical data
}
