% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield_descriptive.R
\name{datashield_descriptive}
\alias{datashield_descriptive}
\title{DataSHIELD descriptive function}
\usage{
datashield_descriptive(dsfunction, opal_connection, save = FALSE, df = "D")
}
\arguments{
\item{dsfunction}{The function you want to run (ds.class, ds.numNA, ds.length)}

\item{opal_connection}{An Opal connection}

\item{save}{if TRUE, the output is saved in the working directory as a csv file}

\item{df}{specifies the  df that was assigned in the login, default is "D"}
}
\value{
A table with the output of the function, if more than one study, they are joined in one table
}
\description{
DataSHIELD descriptive function
}
\examples{
datashield_descriptive(ds.class, opals)
}
