% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield_histogram.R
\name{datashield_histogram}
\alias{datashield_histogram}
\title{DataSHIELD helper function for histograms}
\usage{
datashield_histogram(datasources = NULL, method = "deterministic")
}
\arguments{
\item{datasources}{An Opal connection}

\item{method}{Type of method for building the histogram}
}
\value{
It creates a directory with subdirectories of each study, containing histograms of all variables
}
\description{
DataSHIELD helper function for histograms
}
