% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield_summary.R
\name{datashield_summary}
\alias{datashield_summary}
\title{It runs the datashield function ds.summary()and calculates the standard deviation in all numeric and integer variables of each OpalConnection}
\usage{
datashield_summary(opal_connection, save_summary = FALSE)
}
\arguments{
\item{opal_connection}{An Opal connection}

\item{save_summary}{if TRUE, the output is saved in the working directory as a csv file}
}
\value{
a list with the summary of each variable
}
\description{
It runs the datashield function ds.summary()and calculates the standard deviation in all numeric and integer variables of each OpalConnection
}
\examples{
datashield_summary(opals)
}
